Checkpoint engine before ownership migration

+++ OpenGL/src/scene/StaticMeshObject.cpp
@@ -1,10 +1,10 @@
 #include "StaticMeshObject.h"
 
 StaticMeshObject::StaticMeshObject(StaticMesh* gameObject,
-	double x,
-	double y,
-	double z
-) : x(x), y(y), z(z), rotX(0), rotY(0), rotZ(0), scaleX(1), scaleY(1), scaleZ(1), staticGeometry(gameObject)
+	float32 x,
+	float32 y,
+	float32 z
+) : x(x), y(y), z(z), rotX(0.0f), rotY(0.0f), rotZ(0.0f), scaleX(1.0f), scaleY(1.0f), scaleZ(1.0f), staticGeometry(gameObject)
 {
 }
 
@@ -22,24 +22,24 @@ const StaticMesh* const StaticMeshObject::getGameObject() const {
 	return this->staticGeometry;
 }
 
-void StaticMeshObject::setPosition(float x, float y, float z) {
+void StaticMeshObject::setPosition(float32 x, float32 y, float32 z) {
 	this->x = x;
 	this->y = y;
 	this->z = z;
 }
 
-void StaticMeshObject::setRotation(float rotX, float rotY, float rotZ) {
+void StaticMeshObject::setRotation(float32 rotX, float32 rotY, float32 rotZ) {
 	this->rotX = rotX;
 	this->rotY = rotY;
 	this->rotZ = rotZ;
 }
 
-void StaticMeshObject::setScale(float scaleX, float scaleY, float scaleZ) {
+void StaticMeshObject::setScale(float32 scaleX, float32 scaleY, float32 scaleZ) {
 	this->scaleX = scaleX;
 	this->scaleY = scaleY;
 	this->scaleZ = scaleZ;
 }
 
-void StaticMeshObject::setScale(float xyz) {
+void StaticMeshObject::setScale(float32 xyz) {
 	this->setScale(xyz, xyz, xyz);
 }
